LayoutTests/js/basic-spread-expected.txt

This test checks the behavior of the spread construct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS passedThis is o
PASS args[0] is 1
PASS args[1] is undefined
PASS args[2] is null
PASS args[3] is 4
PASS passedThis is o
PASS args[0] is 1
PASS args[1] is undefined
PASS args[2] is null
PASS args[3] is 4
PASS passedThis is o
PASS args[0] is 1
PASS args[1] is undefined
PASS args[2] is null
PASS args[3] is 4
PASS passedThis is o
PASS args[0] is 1
PASS args[1] is undefined
PASS args[2] is null
PASS args[3] is 4
PASS passedThis is o
PASS args[0] is 1
PASS args[1] is undefined
PASS args[2] is null
PASS args[3] is 4
PASS passedThis is o
PASS args[0] is 1
PASS args[1] is undefined
PASS args[2] is null
PASS args[3] is 4
PASS passedThis is o
PASS args[0] is 1
PASS args[1] is undefined
PASS args[2] is null
PASS args[3] is 4
PASS passedThis is o
PASS args[0] is 1
PASS args[1] is undefined
PASS args[2] is null
PASS args[3] is 4
PASS a is [1,2,3]
PASS [...a] is [1,2,3]
PASS [...a] is [1,2,3]
PASS [...a,...[...a]] is [1,2,3,1,2,3]
PASS [,,,...a] is [,,,1,2,3]
PASS [...a,,,].join('|') is [1,2,3,,,].join('|')
PASS [,...a,4] is [,1,2,3,4]
PASS [,...a,,5] is [,1,2,3,,5]
PASS [...a.keys()] is [0,1,2]
PASS [...a.entries()].join('|') is [[0,1],[1,2],[2,3]].join('|')
PASS [...a] is [1,2,3]
PASS successfullyParsed is true

TEST COMPLETE
